// src/payload-types.ts
export interface Model {
  id: string;
  name: string;
  description: string;
  about: {
    root: {
      type: string;
      children: {
        type: string;
        version: number;
        [k: string]: unknown;
      }[];
      direction: ('ltr' | 'rtl') | null;
      format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
      indent: number;
      version: number;
    };
    [k: string]: unknown;
  };
  type: 'text' | 'image';
  imageinput: boolean;
  provider: 'groq' | 'openai' | 'replicate';
  settings: {
    name: string;
    type: 'integer' | 'string' | 'boolean' | 'select';
    description: string;
    options?: string[] | null;
    allowMultiple: boolean;
    id?: string | null;
  }[];
  systemprompt: boolean;
  negativeprompt: boolean;
  prodkeys?: {
    groq?: {
      modelname: string;
      apikey: string;
    };
    replicate?: {
      modelname: string;
      apitoken: string;
    };
    openai?: {
      modelname: string;
      apikey: string;
    };
  };
  testkeys?: {
    groq?: {
      modelname: string;
      apikey: string;
    };
    replicate?: {
      modelname: string;
      apitoken: string;
    };
    openai?: {
      modelname: string;
      apikey: string;
    };
  };
  cost: number;
  commissionapplicable: boolean;
  commission?: number | null;
  examples: {
    example: string;
    id?: string | null;
  }[];
  updatedAt: string;
  createdAt: string;
}